2.0.6
** Breaking changes **
Former Span parameters to implementations were good in many ways and was a step towards reducing heap allocations in time critical code paths as much as possible. However, since Span<T> is a ref struct, it was not possible to send it on to asynchronous APIs or lambda functions and similar.

So, in this version, all ReadOnlySpan<T> parameters have been changed to ReadOnlyDokanMemory<T> and all Span<T> parameters have been changed to DokanMemory<T>. This needs changes in implementing classes. But it should be easy, the new structs have Span properties that return Span<T> that could be used like before. But furthermore, the new structs are not ref structs so they can be passed to lambda functions (and Span property called inside them if a Span is needed) and there is also a GetMemoryManager() method that returns a MemoryManager object that can be used to get a Memory<T> for use with asynchronous APIs. There is also an Address property that returns the unmanaged memory address supplied by Dokan native library.